corrected some minor error
