+++ home/index.html
@@ -5,11 +5,11 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Home Page</title>
-    <link rel="icon" href="images/logo.png" type="image/png">
-    <link rel="stylesheet" href="css folder/homepageheader.css">
-    <link rel="stylesheet" href="css folder/homepagebody.css">
-    <link rel="stylesheet" href="css folder/divinegracegeneral.css">
-    <link rel="stylesheet" href="css folder/divinegracemain.css">
+    <link rel="icon" href="/Divine-Grace-Upgrade/images/logo.png" type="image/png">
+    <link rel="stylesheet" href="/Divine-Grace-Upgrade/css folder/homepageheader.css">
+    <link rel="stylesheet" href="/Divine-Grace-Upgrade/css folder/homepagebody.css">
+    <link rel="stylesheet" href="/Divine-Grace-Upgrade/css folder/divinegracegeneral.css">
+    <link rel="stylesheet" href="/Divine-Grace-Upgrade/css folder/divinegracemain.css">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Agu+Display&display=swap" rel="stylesheet">
     <link href="https://fonts.googleapis.com/css2?family=Nunito+Sans&family=Open+Sans&display=swap" rel="stylesheet">
@@ -18,10 +18,10 @@
     <header class="main-contents" id="header">
       <div class="right-section">
         <button class="hambuger-button js-hambuger-button">
-          <img class="hambuger js-hambuger" src="images/hamburger-menu.svg" alt="Hamburger Menu">
+          <img class="hambuger js-hambuger" src="/Divine-Grace-Upgrade/images/hamburger-menu.svg" alt="Hamburger Menu">
         </button>
         <div class="logo-div">
-          <img id="logo-img" src="images/logo.png" alt="">
+          <img id="logo-img" src="/Divine-Grace-Upgrade/images/logo.png" alt="">
           <span id="digunec-logo">Divine Grace Unec</span>
         </div>
         <div id="divider"></div>
@@ -29,27 +29,27 @@
       </div>
       <div class="left-section">
         <button>
-          <img id="notification" src="images/notifications.svg" alt="">
+          <img id="notification" src="/Divine-Grace-Upgrade/images/notifications.svg" alt="">
         </button>
-        <img id="profile-image" src="images/th-2238308759" alt="">
+        <img id="profile-image" src="/Divine-Grace-Upgrade/images/th-2238308759" alt="">
       </div>
     </header>
    <div>
     <nav class="nav-menu hidden" id="nav-menue">
         <div class="logo-nav">
             <div class="logo-div">
-                <img id="logo-img" src="images/logo.png" alt="">
+                <img id="logo-img" src="/Divine-Grace-Upgrade/images/logo.png" alt="">
                 <span id="digunec-logo">Divine Grace Unec</span>
               </div>
               <button class="exit-butt js-exit-butt">
-                <img src="images/exit-button-png-close-close-button-exit-quit-x-icon-512.png" alt="Exit">
+                <img src="/Divine-Grace-Upgrade/images/exit-button-png-close-close-button-exit-quit-x-icon-512.png" alt="Exit">
               </button>
         </div>
         <nav class="nav">
-            <li class="menu-item1"><a href="homepage.html">Dashboard</a></li>
-            <li><a href="prayerequest.html">Submit a Prayer Request</a></li>
+            <li class="menu-item1"><a href="/Divine-Grace-Upgrade/home">Dashboard</a></li>
+            <li><a href="/Divine-Grace-Upgrade/prayerequest">Submit a Prayer Request</a></li>
             <li><a href="https://wa.me/2348141787294">Connect with Us</a></li>
-            <li><a href="lstsregistrationpage.html">Register for LSTS</a><li>
+            <li><a href="/Divine-Grace-Upgrade/lstsregistrationpage">Register for LSTS</a><li>
             <li><button class="signout-button js-sign-out-button">Sign Out</button></li>
           </nav>
       </nav>
@@ -62,12 +62,12 @@
                   <section class="main-body-section">
                     <section class="left-body-section">
                       <button class="exit-button js-exit-button">
-                        <img src="images/exit-button-png-close-close-button-exit-quit-x-icon-512.png" alt="">
+                        <img src="/Divine-Grace-Upgrade/images/exit-button-png-close-close-button-exit-quit-x-icon-512.png" alt="">
                       </button>
                       <div class="main-left-div">
                         <div class="profile-imageH-div">
                           <div>
-                            <img class="profile-imageH" id="user-profile-image" src="images/th-2238308759" alt="">
+                            <img class="profile-imageH" id="user-profile-image" src="/Divine-Grace-Upgrade/images/th-2238308759" alt="">
                           </div>
                           <p class="profile-name" id="user-name"></p>
                           <p class="department">Department in the church: <span class="department-span" id="user-department"></span></p>
@@ -105,7 +105,7 @@
                           <p class="p">ANNOUNCEMENT</p>
                         </div>
                         <div id="annoucement-image">
-                          <img class="annoucement-img" src="images/annocement_image-removebg-preview.png" alt="">
+                          <img class="annoucement-img" src="/Divine-Grace-Upgrade/images/annocement_image-removebg-preview.png" alt="">
                           <p class="annoucement-p">Announcements and updates will appear here</p>
                         </div>
                       </div>
@@ -113,13 +113,13 @@
                   </section>
                   <footer>
                     <div class="footer-logo">
-                      <img id="logo-img" src="images/logo.png" alt="">
+                      <img id="logo-img" src="/Divine-Grace-Upgrade/images/logo.png" alt="">
                       <span id="digunec-logo">DIGUNEC</span>
                     </div>
                     <div class="social-div">
-                      <img src="images/facebook-logo-2428.png" alt="">
-                      <img src="images/instagram-logo-8869.png" alt="">
-                      <img src="images/twitter-x-blue-logo-round-20859.png" alt="">
+                      <img src="/Divine-Grace-Upgrade/images/facebook-logo-2428.png" alt="">
+                      <img src="/Divine-Grace-Upgrade/images/instagram-logo-8869.png" alt="">
+                      <img src="/Divine-Grace-Upgrade/images/twitter-x-blue-logo-round-20859.png" alt="">
                     </div>
                     <div class="copyright-div">
                       <p class="copyright">&copy;copyright <b>DIGUNEC</b>&#46; All rights reserved</p>
@@ -142,6 +142,6 @@
         
     </div>
 
-    <script type="module" src="scripts/homepage.js"></script>
+    <script type="module" src="/Divine-Grace-Upgrade/scripts/homepage.js"></script>
 </body>
 </html>
