+++ homepage/index.html
@@ -4,12 +4,12 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>DIVINE GRACE UNEC</title>
-    <link rel="stylesheet" href="css folder/divinegraceheader.css">
-    <link rel="stylesheet" href="css folder/divinegracegeneral.css">
-    <link rel="stylesheet" href="css folder/divinegracemain.css">
-    <link rel="stylesheet" href="css folder/welcomepage.css">
+    <link rel="stylesheet" href="/Divine-Grace-Upgrade/css folder/divinegraceheader.css">
+    <link rel="stylesheet" href="/Divine-Grace-Upgrade/css folder/divinegracegeneral.css">
+    <link rel="stylesheet" href="/Divine-Grace-Upgrade/css folder/divinegracemain.css">
+    <link rel="stylesheet" href="/Divine-Grace-Upgrade/css folder/welcomepage.css">
     <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="icon" href="images/logo.png" type="image/png">
+    <link rel="icon" href="/Divine-Grace-Upgrade/images/logo.png" type="image/png">
 <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
 <link href="https://fonts.googleapis.com/css2?family=Nunito+Sans:ital,opsz,wght@0,6..12,200..1000;1,6..12,200..1000&family=Open+Sans:ital,wght@0,300..800;1,300..800&display=swap" rel="stylesheet">
 </head>
@@ -17,7 +17,7 @@
     <section class="header-section">
         <header class="header-div" id="header">
             <div class="add">
-                <img class="logo-img" src="images/logo.png" alt="">
+                <img class="logo-img" src="/Divine-Grace-Upgrade/images/logo.png" alt="">
                 <span class="digunec">
                     DIG UNEC
                 </span>
@@ -44,7 +44,7 @@
                 </p>
         </div>
         <div class="welcome-flex-image-div">
-            <img class="welcome-page-image" src="images/1732910927627.jpg" alt="" class="welcome-flex-image">
+            <img class="welcome-page-image" src="/Divine-Grace-Upgrade/images/1732910927627.jpg" alt="" class="welcome-flex-image">
         </div>
       </div>
         <div class="who">WHO WE ARE</div>
@@ -63,7 +63,7 @@
         <div class="ourvalue-grid">
             <div class="colour-grid">
                 <div class="snappic1-div" >
-                    <img src="images/first image.jpg" alt="" class="snappic1">
+                    <img src="/Divine-Grace-Upgrade/images/first image.jpg" alt="" class="snappic1">
                 </div>
                 <div>
                     <p class="soul-p">Soul Winning</p>
@@ -73,7 +73,7 @@
             </div>
             <div class="colour-grid">
                 <div class="snappic1-div">
-                    <img src="images/snappic.jpg" alt="" class="snappic2">
+                    <img src="/Divine-Grace-Upgrade/images/snappic.jpg" alt="" class="snappic2">
                 </div>
                 <div>
                     <p class="vision-p">Vision </p>
@@ -84,7 +84,7 @@
         <div class="ourvalue-grid">
             <div class="colour-grid">
                 <div class="snappic1-div" >
-                    <img src="images/first image.jpg" alt="" class="snappic1">
+                    <img src="/Divine-Grace-Upgrade/images/first image.jpg" alt="" class="snappic1">
                 </div>
                 <div>
                     <p class="soul-p">Prayer</p>
@@ -97,7 +97,7 @@
             </div>
             <div class="colour-grid">
                 <div class="snappic1-div">
-                    <img src="images/snappic.jpg" alt="" class="snappic2">
+                    <img src="/Divine-Grace-Upgrade/images/snappic.jpg" alt="" class="snappic2">
                 </div>
                 <div>
                     <p class="vision-p">Rightfully dividing the word of God</p>
@@ -113,23 +113,23 @@
    <div class="service-section">
     <div class="Service">SERVICES</div>
     <div class="Service-time">
-        Tuesday (Mid-week Service) <br> Time <img src="images/pngaaa.com-473864.png" alt="" class="time-image">: 5:00pm
+        Tuesday (Mid-week Service) <br> Time <img src="/Divine-Grace-Upgrade/images/pngaaa.com-473864.png" alt="" class="time-image">: 5:00pm
     </div>
     <div class="Service-time">
         Friday (LSTS)
-        <a class="lsts-anchor" href="registerlogin.html">Click To Register</a> <br>Time <img src="images/pngaaa.com-473864.png" alt="" class="time-image">: 5:00pm
+        <a class="lsts-anchor" href="registerlogin.html">Click To Register</a> <br>Time <img src="/Divine-Grace-Upgrade/images/pngaaa.com-473864.png" alt="" class="time-image">: 5:00pm
     </div>
     <div class="Service-time">
-        Sunday Service <br> Time <img src="images/pngaaa.com-473864.png" alt="" class="time-image">: 8:00am
+        Sunday Service <br> Time <img src="/Divine-Grace-Upgrade/images/pngaaa.com-473864.png" alt="" class="time-image">: 8:00am
     </div>
     <div>
         <div class="Service-time">
-            Venue<img src="images/Location_icon_from_Noun_Project.png" alt="" class="time-image">: Ulasi Building opposite Unec Library
+            Venue<img src="/Divine-Grace-Upgrade/images/Location_icon_from_Noun_Project.png" alt="" class="time-image">: Ulasi Building opposite Unec Library
         </div>
     </div>
     <div class="Service-time">
         <p class="callus">For more enquires call us
-            <img src="images/kisspng-computer-icons-telephone-call-5af1f6d96f06b6.4041455015258068094548.png" alt="" class="time-image"></p>
+            <img src="/Divine-Grace-Upgrade/images/kisspng-computer-icons-telephone-call-5af1f6d96f06b6.4041455015258068094548.png" alt="" class="time-image"></p>
        <p class="callus">+2348126640203 <br>+2349060697685 
            <br>+2348139952179 <br>+2347041629192 <br> Email: divinegraceunec@gmail.com</p>
     </div>
@@ -138,15 +138,15 @@
 </section>
     <footer>
         <div class="footer-logo">
-            <img class="logo-imgg" src="images/logo.png" alt="">
+            <img class="logo-imgg" src="/Divine-Grace-Upgrade/images/logo.png" alt="">
             <span class="digunecc">
                 DIGUNEC
             </span>
         </div>
         <div class="social-div">
-            <img src="images/facebook-logo-2428.png" alt="">
-            <img src="images/instagram-logo-8869.png" alt="">
-            <img src="images/twitter-x-blue-logo-round-20859.png" alt="">
+            <img src="/Divine-Grace-Upgrade/images/facebook-logo-2428.png" alt="">
+            <img src="/Divine-Grace-Upgrade/images/instagram-logo-8869.png" alt="">
+            <img src="/Divine-Grace-Upgrade/images/twitter-x-blue-logo-round-20859.png" alt="">
         </div>
         <div class="copyright-div">
             <p class="copyright">&copy;copyright <b>DIGUNEC</b>&#46; All right reserved</p>
@@ -154,6 +154,6 @@
         </div>
     </footer>
 
-    <script type="module" src="scripts/divinegracegeneral.js"></script>
+    <script type="module" src="/Divine-Grace-Upgrade/scripts/divinegracegeneral.js"></script>
 </body>
 </html>
