+++ lstsregistrationpage/index.html
@@ -4,14 +4,14 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>LSTS REGISTRATION PAGE</title>
-    <link rel="stylesheet" href="css folder/lsts.css">
-    <link rel="icon" href="images/logo.png" type="image/png">
-    <link rel="stylesheet" href="css folder/homepageheader.css">
-    <link rel="stylesheet" href="css folder/prayerrequest.css">
-    <link rel="stylesheet" href="css folder/prayerrequest.css">
-    <link rel="stylesheet" href="css folder/homepagebody.css">
-    <link rel="stylesheet" href="css folder/divinegracegeneral.css">
-    <link rel="stylesheet" href="css folder/divinegracemain.css">
+    <link rel="stylesheet" href="/Divine-Grace-Upgrade/css folder/lsts.css">
+    <link rel="icon" href="/Divine-Grace-Upgrade/images/logo.png" type="image/png">
+    <link rel="stylesheet" href="/Divine-Grace-Upgrade/css folder/homepageheader.css">
+    <link rel="stylesheet" href="/Divine-Grace-Upgrade/css folder/prayerrequest.css">
+    <link rel="stylesheet" href="/Divine-Grace-Upgrade/css folder/prayerrequest.css">
+    <link rel="stylesheet" href="/Divine-Grace-Upgrade/css folder/homepagebody.css">
+    <link rel="stylesheet" href="/Divine-Grace-Upgrade/css folder/divinegracegeneral.css">
+    <link rel="stylesheet" href="/Divine-Grace-Upgrade/css folder/divinegracemain.css">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Agu+Display&display=swap" rel="stylesheet">
     <link href="https://fonts.googleapis.com/css2?family=Nunito+Sans&family=Open+Sans&display=swap" rel="stylesheet">
@@ -22,10 +22,10 @@
     <header class="main-contents" id="header">
         <div class="right-section">
           <button class="hambuger-button js-hambuger-button">
-            <img class="hambuger js-hambuger" src="images/hamburger-menu.svg" alt="Hamburger Menu">
+            <img class="hambuger js-hambuger" src="/Divine-Grace-Upgrade/images/hamburger-menu.svg" alt="Hamburger Menu">
           </button>
           <div class="logo-div">
-            <img id="logo-img" src="images/logo.png" alt="">
+            <img id="logo-img" src="/Divine-Grace-Upgrade/images/logo.png" alt="">
             <span id="digunec-logo">Divine Grace Unec</span>
           </div>
           <div id="divider"></div>
@@ -33,27 +33,27 @@
         </div>
         <div class="left-section">
           <button>
-            <img id="notification" src="images/notifications.svg" alt="">
+            <img id="notification" src="/Divine-Grace-Upgrade/images/notifications.svg" alt="">
           </button>
-          <img id="profile-image" src="images/th-2238308759" alt="">
+          <img id="profile-image" src="/Divine-Grace-Upgrade/images/th-2238308759" alt="">
         </div>
       </header>
      <div>
       <nav class="nav-menu hidden">
           <div class="logo-nav">
               <div class="logo-div">
-                  <img id="logo-img" src="images/logo.png" alt="">
+                  <img id="logo-img" src="/Divine-Grace-Upgrade/images/logo.png" alt="">
                   <span id="digunec-logo">Divine Grace Unec</span>
                 </div>
                 <button class="exit-butt js-exit-butt">
-                  <img src="images/exit-button-png-close-close-button-exit-quit-x-icon-512.png" alt="Exit">
+                  <img src="/Divine-Grace-Upgrade/images/exit-button-png-close-close-button-exit-quit-x-icon-512.png" alt="Exit">
                 </button>
           </div>
           <nav class="nav">
-            <li><a href="homepage.html">Dashboard</a></li>
-            <li><a href="prayerequest.html">Submit a Prayer Request</a></li>
+            <li><a href="/Divine-Grace-Upgrade/home">Dashboard</a></li>
+            <li><a href="/Divine-Grace-Upgrade/prayerequest">Submit a Prayer Request</a></li>
             <li><a href="https://wa.me/2348141787294">Connect with Us</a></li>
-            <li class="menu-item2"><a href="lstsregistrationpage.html">Register for LSTS</a><li>
+            <li class="menu-item2"><a href="/Divine-Grace-Upgrade/lstsregistrationpage">Register for LSTS</a><li>
             <li><button class="signout-button js-sign-out-button">Sign Out</button></li>
           </nav>
         </nav>
@@ -101,6 +101,6 @@
         <button class="submit-button" type="submit">Submit</button>
     </form>
 </section>
-<script type="module" src="scripts/lsts.js"></script>
+<script type="module" src="/Divine-Grace-Upgrade/scripts/lsts.js"></script>
 </body>
 </html>
