+++ registerlogin/index.html
@@ -4,10 +4,10 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>DIVINE GRACE UNEC</title>
-    <link rel="stylesheet" href="css folder/divinegracemain.css">
-    <link rel="stylesheet" href="css folder/divinegraceheader.css">
-    <link rel="stylesheet" href="css folder/registerlogin.css">
-    <link rel="icon" href="images/logo.png" type="image/png">
+    <link rel="stylesheet" href="/Divine-Grace-Upgrade/css folder/divinegracemain.css">
+    <link rel="stylesheet" href="/Divine-Grace-Upgrade/css folder/divinegraceheader.css">
+    <link rel="stylesheet" href="/Divine-Grace-Upgrade/css folder/registerlogin.css">
+    <link rel="icon" href="/Divine-Grace-Upgrade/images/logo.png" type="image/png">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
 <link href="https://fonts.googleapis.com/css2?family=Nunito+Sans:ital,opsz,wght@0,6..12,200..1000;1,6..12,200..1000&family=Open+Sans:ital,wght@0,300..800;1,300..800&display=swap" rel="stylesheet">
 </head>
@@ -73,21 +73,21 @@
 
     <footer>
         <div class="footer-logo">
-            <img class="logo-imgg" src="images/logo.png" alt="">
+            <img class="logo-imgg" src="/Divine-Grace-Upgrade/images/logo.png" alt="">
             <span class="digunecc">
                 DIGUNEC
             </span>
         </div>
         <div class="social-div">
-            <img src="images/facebook-logo-2428.png" alt="">
-            <img src="images/instagram-logo-8869.png" alt="">
-            <img src="images/twitter-x-blue-logo-round-20859.png" alt="">
+            <img src="/Divine-Grace-Upgrade/images/facebook-logo-2428.png" alt="">
+            <img src="/Divine-Grace-Upgrade/images/instagram-logo-8869.png" alt="">
+            <img src="/Divine-Grace-Upgrade/images/twitter-x-blue-logo-round-20859.png" alt="">
         </div>
         <div class="copyright-div">
             <p class="copyright">&copy;copyright <b>DIGUNEC</b>&#46; All right reserved</p>
             <p class="design"><i>Designed by Soft Nation</i></p>
         </div>
     </footer>
-    <script type="module" src="scripts/registerlogin.js"></script>
+    <script type="module" src="/Divine-Grace-Upgrade/scripts/registerlogin.js"></script>
 </body>
 </html>
