+++ scripts/divinegracegeneral.js
@@ -1,3 +1,3 @@
 document.querySelector('.js-login-button').addEventListener('click', ()=>{
-    window.location.href = 'registerlogin.html';
+    window.location.href = '/Divine-Grace-Upgrade/registerlogin';
 });
