+++ scripts/lsts.js
@@ -1,6 +1,8 @@
-import { renderHeader } from "./homepage.js";
+import { renderHeader } from "./utils.js";
+import { wireLogout } from "./utils.js";
 
 renderHeader();
+wireLogout();
