+++ scripts/prayerrequest.js
@@ -1,3 +1,5 @@
-import { renderHeader } from "./homepage.js";
+import { renderHeader } from "./utils.js";
+import { wireLogout } from "./utils.js";
 
 renderHeader();
+wireLogout();
